Refactor(frontend): Move ThemeToggler own file
Better to separate it as a singular component, should probably have a
component/ folder that stores all non-view components that are essential
to views.
