+++ frontend/views/Settings.tsx
@@ -3,32 +3,35 @@ import { View } from "react-native";
 
 import { StackNavigationProp } from "@react-navigation/stack";
 import { RouteProp } from "@react-navigation/native";
-import { Checkbox, List, IconButton, Colors } from "react-native-paper";
+import { Divider } from "react-native-paper";
 import { bindActionCreators, Dispatch } from "redux";
 import { connect } from "react-redux";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 
 import { Routes } from "../routes";
-import { Store, updateTheme, updateUser } from "../redux";
-import { ThemeStorage, ThemeType } from "../theme";
+import { Store, updateUser } from "../redux";
+import ThemeToggler from "./ThemeToggler";
 
 type Props = {
   navigation: StackNavigationProp<Routes, "Settings">;
   route: RouteProp<Routes, "Settings">;
   token: string;
   updateUser: (user: Store["user"]) => void;
-  updateTheme: (theme: ThemeType) => void;
 };
 
 function Settings(props: Props): React.ReactElement {
-  return <ThemeToggler onChange={props.updateTheme} />;
+  return (
+    <View>
+      <ThemeToggler />
+      <Divider />
+    </View>
+  );
 }
 
 const mapDispatchToProps = (dispatch: Dispatch) =>
   bindActionCreators(
     {
       updateUser,
-      updateTheme,
     },
     dispatch
   );
@@ -47,65 +50,3 @@ export default connect(
   }),
   mapDispatchToProps
 )(Settings);
-
-type DarkThemeToggler = {
-  onChange: (type: ThemeType) => void;
-  theme: ThemeType;
-};
-
-function DarkThemeToggler(props: DarkThemeToggler): React.ReactElement {
-  const { onChange, theme } = props;
-
-  useEffect(() => {
-    const getTheme = async () => {
-      let theme: ThemeType;
-      try {
-        theme = (await AsyncStorage.getItem(ThemeStorage)) as ThemeType;
-      } catch (err) {
-        theme = "Light";
-      }
-
-      if (!theme) {
-        theme = "Light";
-      }
-
-      updateTheme(theme);
-    };
-
-    getTheme();
-  }, []);
-
-  const handleChange = async () => {
-    const newTheme = theme === "Dark" ? "Light" : "Dark";
-    onChange(newTheme);
-    try {
-      await AsyncStorage.setItem(ThemeStorage, newTheme);
-    } catch (err) {
-      console.log("failed to set theme");
-    }
-  };
-
-  return (
-    <View>
-      <List.Item
-        title="Dark Theme"
-        onPress={handleChange}
-        left={() => <IconButton icon="brightness-6" />}
-        right={() => (
-          <Checkbox
-            status={theme === "Dark" ? "checked" : "unchecked"}
-            color={Colors.grey800}
-          />
-        )}
-      />
-    </View>
-  );
-}
-
-const ThemeToggler = connect(
-  (state: Store, ownProps: { onChange: (theme: ThemeType) => void }) => ({
-    onChange: ownProps.onChange,
-    theme: state.theme,
-  }),
-  null
-)(DarkThemeToggler);
